# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Search Social Media profiles and estimate salary"
  version: "1.0.0"
  title: "DataChallenge API"
host: "localhost:9090"
basePath: "/api/v1"
schemes:
- "http"
produces:
- "application/json"
- "application/xml"
paths:
...
  /users/authenticate/{provider}:
    get:
      tags:
      - "Authentication"
      summary: "Perform login"
      description: "Perform Basic login"
      operationId: "usersAuthenticateGET"
      parameters:
      - name: "provider"
        in: "path"
        description: "provider used for login"
        required: true
        type: "string"
        enum:
        - "basic"
        - "xing"
        - "linkedIn"
      - name: "Authorization"
        in: "header"
        description: "Basic authorization credentials"
        required: false
        type: "string"
      responses:
        204:
          description: "Auth ok. User is logged in"
        403:
          description: "Invalid credentials"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Authentication"
  /users/authenticate/{provider}/callback:
    get:
      tags:
      - "Authentication"
      summary: "OAuth Callback"
      description: "Internal callback route for oauth processes"
      operationId: "usersAuthenticateProviderCallbackGET"
      parameters:
      - name: "provider"
        in: "path"
        description: "provider used for login"
        required: true
        type: "string"
        enum:
        - "xing"
        - "linkedIn"
      responses:
        204:
          description: "Callback received"
      x-swagger-router-controller: "Authentication"
...
definitions:
...
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"
    xml:
      name: "error"